requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiohttp-jinja2==1.6
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
certifi==2024.8.30
click==8.1.7
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
Jinja2==3.1.4
linkify-it-py==2.0.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdit-py-plugins==0.4.2
mdurl==0.1.2
msgpack==1.1.0
multidict==6.1.0
platformdirs==4.3.6
portion==2.6.0
propcache==0.2.1
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
python-dotenv==1.0.1
rich==13.9.3
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.36
sqlmodel==0.0.22
textual==1.0.0
textual-dev==1.7.0
textual-serve==1.1.1
typing_extensions==4.12.2
uc-micro-py==1.0.3
yarl==1.18.3
